requirements.txt
libvirt-python
click
Jinja2
fabric
PyYaml

marshmallow
marshmallow-dataclass
typeguard
rich
